feat: ✨ add renderview job to sync queue
