+++ packages/x6-core/src/index.ts
@@ -1,5 +1,9 @@
+import * as Registry from './registry'
+
 export * from './common'
 export * from './model'
 export * from './view'
 export * from './renderer'
 export * from './util'
+
+export { Registry }
