--- packages/x6-core/src/schedule/index.ts
@@ -1,1 +1,0 @@
-export * from './schedule'
